robot_framework_test/requirements.txt
robotframework==3.1.2
selenium==3.14.1
tk==0.1.0
